WIP: Change robot movement to tracked
This is instead of the "pushed" movement model. The aim is to make
the robot behave more realistically and therefore make the gap
between real world and simulation narrower.

This commit is work in progress and should be amended before merging
to main-line. Proto code does not support the new model yet.